+++ Attribute/Reflection/ReflectionMember.php
@@ -96,4 +96,19 @@ class ReflectionMember
     {
         return $this->member instanceof \ReflectionParameter ? 'parameter' : 'property';
     }
+
+    public function isBackedEnumType(): bool
+    {
+        return $this->member->getType() instanceof \ReflectionNamedType && is_subclass_of($this->member->getType()->getName(), \BackedEnum::class);
+    }
+
+    public function isParameter(): bool
+    {
+        return $this->member instanceof \ReflectionParameter;
+    }
+
+    public function isProperty(): bool
+    {
+        return $this->member instanceof \ReflectionProperty;
+    }
 }