+++ Tests/Fixtures/InvokableWithInputTestCommand.php
@@ -13,6 +13,7 @@ namespace Symfony\Component\Console\Tests\Fixtures;
 
 use Symfony\Component\Console\Attribute\Argument;
 use Symfony\Component\Console\Attribute\AsCommand;
+use Symfony\Component\Console\Attribute\Interact;
 use Symfony\Component\Console\Attribute\MapInput;
 use Symfony\Component\Console\Attribute\Option;
 use Symfony\Component\Console\Command\Command;
@@ -21,6 +22,12 @@ use Symfony\Component\Console\Style\SymfonyStyle;
 #[AsCommand('invokable:input:test')]
 class InvokableWithInputTestCommand
 {
+    #[Interact]
+    public function interact(SymfonyStyle $io, #[MapInput] UserDto $user): void
+    {
+        $user->email ??= 'user.interactive@command.com';
+    }
+
     public function __invoke(SymfonyStyle $io, #[MapInput] UserDto $user): int
     {
         $io->writeln($user->name);
@@ -58,6 +65,12 @@ final class UserDto
 
     #[Option]
     public UserStatus $status = UserStatus::Unverified;
+
+    #[Interact]
+    public function interact(SymfonyStyle $io): void
+    {
+        $this->password ??= 'user-dto-interactive-password';
+    }
 }
 
 final class UserGroupDto